+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.20.0
+- Merchant Connection update to v2.2.2, see: https://github.com/BlinkReceipt/br_account_linking/blob/master/CHANGELOG.md
+- Ability to link IMAP account with existing App Password
+- Improve logging of Amazon errors
+
 # 2.19.0
 - Merchant Connection update to v2.2.0, see: https://github.com/BlinkReceipt/br_account_linking/blob/master/CHANGELOG.md
 - Fix bug in prod intel for Merchant Connections